% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleInput.R
\name{circleInput}
\alias{circleInput}
\title{A Shiny Circle Input}
\usage{
circleInput(inputId)
}
\arguments{
\item{inputId}{The input id}
}
\value{
the color of the selected circle
}
\description{
A Shiny Circle Input
}
